requirements.txt
Flask>=2.0
requests
pdfminer.six
python-dotenv
openai
